pyproject.toml
[tool.poetry]
name = "doc-comments-ai"
version = "0.1.18"
description = ""
authors = ["fynnfluegge <fynnfluegge@gmx.de>"]
readme = "README.md"
packages = [{ include = "doc_comments_ai" }]

[tool.poetry.dependencies]
python = "^3.9"
tree-sitter-languages = "^1.7.0"
tree-sitter = "^0.20.1"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
tiktoken = "^0.4.0"
yaspin = "^3.0.0"

[tool.poetry.scripts]
aicomment = "doc_comments_ai.__main__:main"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
